move serve to RunnerBase
